🔨 split tooltip components into more files
